requirements.txt
PyQt5==5.15.10
PyMuPDF==1.23.26
python-pptx==0.6.23